add ThemeUtils with method 'isMapsforgeTheme'
