+++ vtm-android-example/src/org/oscim/android/filepicker/ValidRenderTheme.java
@@ -17,28 +17,19 @@
 package org.oscim.android.filepicker;
 
 import org.oscim.theme.ExternalRenderTheme;
-import org.oscim.theme.IRenderTheme;
 import org.oscim.theme.ThemeFile;
-import org.oscim.theme.ThemeLoader;
+import org.oscim.theme.ThemeUtils;
 import org.oscim.theme.XmlMapsforgeThemeBuilder;
 import org.oscim.theme.XmlThemeBuilder;
 import org.oscim.tiling.TileSource.OpenResult;
-import org.xml.sax.Attributes;
 import org.xml.sax.InputSource;
-import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
 import org.xml.sax.helpers.DefaultHandler;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.concurrent.atomic.AtomicBoolean;
 
-import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParserFactory;
-
-import static android.R.attr.theme;
 
 /**
  * Accepts all valid render theme XML files.
@@ -49,42 +40,10 @@ public final class ValidRenderTheme implements ValidFileFilter {
     @Override
     public boolean accept(File file) {
 
-        final AtomicBoolean isMapsforgeTheme = new AtomicBoolean(false);
-        InputStream is=null;
-        try {
-            is= new FileInputStream(file);
-            SAXParserFactory factory = SAXParserFactory.newInstance();
-            factory.setNamespaceAware(true);
-
-            XMLReader xmlReader = factory.newSAXParser().getXMLReader();
-            xmlReader.setContentHandler(new DefaultHandler() {
-
-                public void startElement(String uri, String localName, String qName, Attributes attributes) throws SAXException {
-
-                    if (localName.equals("rendertheme")) {
-                        isMapsforgeTheme.set(uri.equals("http://mapsforge.org/renderTheme"));
-                        //we have all info's, break parsing
-                        throw new ThemeLoader.SAXTerminatorException();
-                    }
-                }
-            });
-            xmlReader.parse(new InputSource(is));
-        } catch (ThemeLoader.SAXTerminatorException e) {
-            //do nothing;
-        } catch (SAXException | IOException | ParserConfigurationException e) {
-            e.printStackTrace();
-        }
-
-        try {
-            is.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
         try {
             ThemeFile theme = new ExternalRenderTheme(file.getAbsolutePath());
             DefaultHandler renderThemeHandler;
-            if(isMapsforgeTheme.get()) {
+            if(ThemeUtils.isMapsforgeTheme(new FileInputStream(file))) {
                 renderThemeHandler = new XmlMapsforgeThemeBuilder(theme);
             }else{
                 renderThemeHandler = new XmlThemeBuilder(theme);
